Digital_segmentation_BRCA_547_Capstone/requirements.txt
matplotlib==3.5.2
numba==0.55.1
numpy==1.21.5
opencv_python==4.5.5.62
openslide_python==1.1.2
scikit_image==0.19.2
scikit_learn==1.1.1
scikit-image
